requirements.txt
pandas
numpy
scikit-learn
Flask
requests
python-dotenv
